moved logic for sending requests into api folder
